Migration to new website
JCAltTestProcess simplified and splited to more classes - part 1.
